+++ apps/api/urls/branch_urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+
+from ..views import branch_views as views
+
+app_name = 'Branchs endpoints'
+urlpatterns = [
+    path('read/list/<int:pk>', views.getBranchList),
+
+]