+++ apps/core/models.py
@@ -267,7 +267,7 @@ class Branch(models.Model):
     )
 
     def __str__(self):
-        return self.name
+        return f'{self.name} {self.id}'
 
 
 class Telebot(models.Model):